Fix ignoring wl_seat's older than v6
The version we actually need is v2, as we only care about the
name event. Technically, even this is not completely correct,
as we can still use a v1 seat seat perfectly well as long as
we're not trying to find a seat with a specific name, but v2
is sufficiently old to be considered mandatory with reason.

v2 is also the version wl-clipboard requires, for the same
reason.

This also adds a unit test for accepting a v2 seat.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>