Replace `derive_more` with `thiserror`
Commit d9829ccb ("Remove failure dependency") made the case that
`thiserror` could be used as a replacement for the `failure` crate just
as well, but opted to use `derive_more` as it was "already used" by
`wl-clipboard-rs`.  Checking the repo before that commit reveals that
this was an unused dependency lingering in `Cargo.toml`, without any
other references to `derive_more`.

Furthermore `derive_more` is an extremely heavy crate that is causing a
multitude of issues for us ([such as inordinate compile times]) which is
why we deter it from our dependency graph altogether.

[such as inordinate compile times]: https://github.com/gfx-rs/rspirv/pull/218#issue-1039357885