Rewrite for wayland-rs 0.30
wayland-rs 0.30 is a major API redesign. This commit ports
wl-clipboard-rs to use it. As a bonus, it fixes
https://github.com/YaLTeR/wl-clipboard-rs/issues/1: the tests are now
fully robust†. As an extra bonus, it adds several randomized
property-based tests.

Since this is a pretty major rewrite, I also took this opportunity to
move to a more standard rustfmt config (Block indent style, default max
width).

Resolves https://github.com/YaLTeR/wl-clipboard-rs/issues/36.

†: the tests can currently fail due to a race condition in wayland-rs,
which will be automatically fixed when
https://github.com/Smithay/wayland-rs/pull/632 is merged.